Run `also_with_pthreads` decorator and use it more places. NFC
This decorator now also delegates to `require_pthreads()` helper method.

Split out from #26827
